Move model/session base classes to opengever.base.
