+++ opengever/base/hooks.py
@@ -1,5 +1,5 @@
+from opengever.base import model
 from opengever.base.config import INDEXES
-from opengever.core import model
 from opengever.core.catalog import add_catalog_indexes
 from Products.CMFCore.utils import getToolByName
 import logging
