+++ opengever/base/model.py
