+++ opengever/core/upgrade.py
@@ -1,7 +1,7 @@
 from alembic.migration import MigrationContext
 from alembic.operations import Operations
 from ftw.upgrade import UpgradeStep
-from opengever.core.model import create_session
+from opengever.base.model import create_session
 from sqlalchemy import Column
 from sqlalchemy import Integer
 from sqlalchemy import MetaData
