+++ opengever/meeting/model/proposal.py
@@ -1,5 +1,5 @@
+from opengever.base.model import Base
 from opengever.base.oguid import Oguid
-from opengever.core.model import Base
 from opengever.meeting import _
 from opengever.meeting.model.query import ProposalQuery
 from opengever.ogds.base.utils import ogds_service
