+++ opengever/ogds/base/sync/ogds_updater.py
@@ -1,7 +1,7 @@
 from App.config import getConfiguration
 from five import grok
 from ldap import NO_SUCH_OBJECT
-from opengever.core.model import create_session
+from opengever.base.model import create_session
 from opengever.ogds.base.interfaces import ILDAPSearch
 from opengever.ogds.base.interfaces import IOGDSUpdater
 from opengever.ogds.base.sync.import_stamp import set_remote_import_stamp
